requirements.txt
asgiref==3.3.0
coverage==5.3
Django==3.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.4.0
djangorestframework==3.10.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
gunicorn
